requirements.txt
appnope==0.1.2
asgiref==3.3.4
backcall==0.2.0
decorator==5.0.7
Django==3.2.2
django-choices==1.7.0
django-import-export==2.5.0
django-filter==2.4.0
django-nested-admin==3.3.3
djangorestframework==3.12.4
django-utils-six==2.0
gunicorn==20.1.0
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
Markdown==3.3.4
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
